Contellationproblem-2020
Constellation Problem
Three characters {#, *, .} represents a constellation of stars and galaxies in space. Each galaxy is demarcated by # characters. There can be one or many stars in a given galaxy. Stars can only be in the shape of vowels {
A
,
E
,
I
,
O
,
U
A,E,I,O,U}. A collection of * in the shape of the vowels is a star. A star is contained in a 3×3 block. Stars cannot be overlapping. The dot(.) character denotes empty space.

Given 
3
×
N
3×N matrix comprising of { #, *, . } character, find the galaxy and stars within them.

Note: Please pay attention to how vowel A is denoted in a 
3
×
3
3×3 block in the examples section below.

Input Format
Input consists of a single integer 
N
N denoting the number of columns.
Output Format
The output contains vowels (stars) in order of their occurrence within the given galaxy. The galaxy itself is represented by the # character.

Constraints
3
≤
N
≤
10
5
3≤N≤10 
5
 
Sample 1:
Input
Output
18
* . * # * * * # * * * # * * * . * .
* . * # * . * # . * . # * * * * * *
* * * # * * * # * * * # * * * * . *
U#O#I#EA
Explanation:
As it can be seen that the stars make the image of the alphabet U, I, and A.

Sample 2:
Input
Output
12
* . * # . * * * # . * .
* . * # . . * . # * * *
* * * # . * * * # * . *
U#I#A
Explanation:
As it can be seen that the stars make the image of the alphabet U, I, and A.